Introducing MELA angles (https://github.com/cp3-llbb/HHAnalysis/issues/59) Moving utilities to Tools
